requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.0
django-filter==23.4
PyJWT==2.1.0
python-dotenv==1.0.1
requests==2.26.0